requirements.txt
Scrapy
scrapy_crawlera
python-dateutil